requirements.txt
numpy==1.21.0
pandas
scikit-learn
statsmodels
yfinance
fbprophet
keras
tensorflow
joblib
jupyter